Can accept webhooks of create & delete, in addition to push.

+++ tools/gotobranch.sh
@@ -1,5 +1,9 @@
 #!/bin/sh
 branch=${1:-"master"}
+if [ ! `git branch -a | grep "remotes/origin/$branch"` ]; then
+    echo "No branch $branch."
+    exit
+fi
 rm -rf content/*
 rm -rf theme/*
 git checkout -f $branch
